Externalize event sender
This commit splits out the EventSender from the EventDealer, so that
consumers of the event dealer can pass the sender in to other event
systems.  This allows for a cleaner separation of the bounds between the
`rest_api` module and the service modules.

Shutdown is handled via drops of the sender.

Additionally:

- Converts the initial events iterator into a stream that is chained to
the receiving side of the sender's channel.  This allows for the
iteration to be handled in a more asynchronous manner, instead of waiting
for the initial events to be processed before returning the response to
the caller.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>