Capture multiple methods in rest_api::Resource
Update the `rest_api::Resource` struct to represent a single URL with
multiple (method, handler) pairs. Previously, a resource captured only a
single (method, handler) pair for a URL, which doesn't match up with
actix resources and doesn't allow for multiple methods on a given URL.

This change deprecates the `Resource::new` method in favor of the
`build` and `add_method` methods to construct a `Resource`.

Signed-off-by: Logan Seeley <seeley@bitwise.io>