Register AuthManager as disconnect listener
When a peer is disconnected, the authorization manager needs to remove
the old state about the peer, otherwise authorization may fail due to
the auth message ordering.

This uses a channel as the mechanism for passing these messages, and
drains the receiver before operating on old state, in order to avoid
deadlocks.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>