Allow unused variable if health features is not enabled
Do to ordering requirments of the different components of
the Splinter Daemon, the inproc used by the health service
needs to be created at the beggining of the start function,
but then used again later. For it to be available later,
we need to allow the unused variable health_inproc.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>