Send past state events to scabbard subscriber
Updates the scabbard service to catch-up a subscriber if it has missed
any state change events. When the subscriber sends the WebSocket
subscription request, it can specify the ID of the last state change
event it received; if this is specified, any state changes that occurred
after the specified one will be sent to the subscriber. If no ID is
specified with the subscription request, all state change events will be
sent (scabbard assumes that this is a new subscriber that will need to
be caught-up).

Signed-off-by: Logan Seeley <seeley@bitwise.io>