Return iterator from list_nodes
Return an iterator from list_nodes, and move the application of the
limit and skip to the caller.  The count of items has been replaced with
an explicit count function, in order to provide the total count of
results.

This does require that registry implementations may need to supply their
own implementations of the iterator to support more efficient skip
behaviour.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>