Add circuit create and circuit vote subcommands to cli
splinter circuit create will take a file path to a yaml
file that defines the circuit propsal.

splinter circuit vote will take a circuit id and
accept/reject for creating a circuit proposal vote for a
circuit proposal.

Both will take a url to point to the splinterd REST API and
a key path for the private key that should be used to sign the
CircuitManagmentPayloads.

Both will currently fail as unimplemented.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>