Add EventConnection trait
An event connection provides an api for subscribing to state change
events.  Implementations provide the subscription implementation with
their underlying source, how to receive the events off of the underlying
stream, and how to unsubscribe.

This commit also includes implementing these traits on a
SawtoothConnection.

The events returned are currently still Sawtooth Event protobuf
messages, but may change in future commits.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>