Move root protos into libsplinter
Having the protos as a peer of libpslinter was preventing publishing
splinter as a crate, because 'cargo publish' only includes files from
the current working directory.

I considered adding the protos directory from the root using the
include field, but the list of files needs to be exhaustive
so this method was considered too error prone.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>