Update gameroom compose files to generate certs with splinter cli
This removes the use of the --generate-cli splinterd flag and replaces
it with generating certificates using the splinter cli in the splinterd
container. Also adds the --insecure flag that is required when using
generated certs, and updates the sample configs to look for the generated
CA file in the correct location.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>