Create signed CircuitManagentPayload
Add the module action::circuit::payload to provide the
make_signed_payload function.  This function can generate signed
payloads for the content of the CircuitManagentPayload, and return the
bytes that will be sent to a splinter node.

This commit provides support for CreateCircuit payloads only.  Future
payloads will need to implement the traits CircuitAction and
ApplyToEnvelope, in order to construct the final payload.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>