Detect exhausted event iterator state
This commit correctly detects an exhausted iterator in the case where
the event iterator has queried to fill its cache, but the resulting
cache is empty.  Previous to this commit, the cache would re-attempt the
query with the empty string as the input. The empty string is replaced
with a valid Fetch condition and an Exhausted condition.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>