Add update Biome users route
Adds a PUT /biome/users/{id} route to Biome. This endpoint can be used
to update a user's password. It requires a payload consisting of the
user's username, current hashed password, and new hashed password. The
request will fail if no user with that username is found or if the
hashed password does not match for that user.

Signed-off-by: Davey Newhall <newhall@bitwise.io>