Add delete Biome user route
Adds a DELETE /biome/users/{id} endpoint to Biome. This endpoint can be
used to delete a user from the database. It takes a payload consisting
of the username and hashed password associated with the user ID passed
in the request. It will fail if the user doesn't exist or the hashed
password is incorrect.

Signed-off-by: Davey Newhall <newhall@bitwise.io>