Remove panics caused by user input on splinterd startup
This commit moves the parsing of args into a new function
that returns. The function returns a UserError if any
argument is not valid or the splinter daemon fails to
start.

This commit also removes expects and panics. Instead
an error will be returned, logged and the program
will be exited.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>