Add RequestGuard trait
This commit introduces the concept of a "request guard" to the REST API
resource framework.  An implementer of a request guard can determine if
a request should be continued or short-circuited before executing the
full action of the resource.  This is achieved by returning a
Continuation of either Continue, or Terminate, the later of which
contains a response.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>