Log as debug when unknown state change received
When a state change is received for a key that gridd does not recognize,
this should be logged as a debug message rather than an error. It's
perfectly normal for state to change in ways that aren't directly
relevant to the grid daemon.

For sabre-related transactions, no debug message is logged to prevent
extraneous messages.

Signed-off-by: Logan Seeley <seeley@bitwise.io>