Move 2PC coordinator timeout default to services
The two-phase commit consensus engine is too low-level to make a
decision about what a good default for the coordinator timeout is, since
this depends on external factors.

This commit updates 2PC to require an explicit coordinator timeout, and
sets reasonable defaults (30 seconds) for the admin service and scabbard
service individually.

Signed-off-by: Logan Seeley <seeley@bitwise.io>