Connect to circuit members when handling votes
Before, connection would only be established when
a create proposal was submitted. This is not sufficient,
because a connection disconnect while waiting for a vote.

This commit adds checking if a peer exists before processing
a vote payload, similar to a create payload.

This commit also breaks out the peer verification for payloads
into its own method.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>