Wait full time in scabbard batch status check
When checking the status of submitted batches, the scabbard client
should wait up to the full time the user specified. The `wait` parameter
provided in the `/batch_statuses` query string is not guaranteed to be
honored by the server, and it may have a limit. This commit updates the
scabbard client to ensure that it waits the full time for batches to
move out of "pending" status by repeatedly checking batch statuses (if
need be).

Signed-off-by: Logan Seeley <seeley@bitwise.io>