Replace service_directory with get_service
Replaces service_directory with get_service. Most uses of
service_directory are followed by a call to HashMap::get, replacing
service_directory with get_service is meant to improve the API, by
making it possible to query for a service directly as well as a way
of reducing the memory foot print of SplinterState as a clone of the
ServiceDirectory won't be created on each call to get_service.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>