Split Event Reactor shutdown
Split `event::Reactor` shutdown function into two parts: a shutdown
signaler and a wait-for-shutdown function.  This allows the signaling to
be done in a non-blocking model and the wait-for-shutdown to be done as
part of a join process.

The old shutdown method is marked as deprecated as of this release.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>