Signal websocket stop actively
This commit replaces the use of the atomic boolean as a
stop/close/disconnect signal with a sender that takes a union type,
WebSocketClientCmd.  This union type is either a Frame or the stop
signal.

This change makes the shutdown essentially immediate, instead of waiting
for a message from the server (most-likely a ping) before disconnecting
the websocket

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>