Allow waiting for complete batches in scabbard
Update the scabbard state to allow specifying a wait time for batches to
complete when getting batch statuses. An iterator is now returned by
this method that, if a timeout is specified, will only provide completed
batch infos. This change removes the need for the REST API to actively
poll for completed batches.

Signed-off-by: Logan Seeley <seeley@bitwise.io>