Remove from_file from toml module
Removes `from_file` from the toml module which was previously used
in place of the TomlConfig object and only compiled when the config-toml
feature was not present. This function no longer serves a purpose and is
not used in Splinterd.

Also removes the `with_source` method from the PartialConfig module as this
method only used in the `from_file` function removed.

Signed-off-by: Shannyn Telander <telander@bitwise.io>