Remove default value from metadata-encoding argument
The presence of the default value set via clap together with the
requires field was causing the metadata argument to always be required.
As metadata is an optional argument that is not the desired behavior.
Removing the default value set via clap fixes that problem. The default
value for metadata-encoding is enforced by the circuit create action.

Signed-off-by: Eloá Franca Verona <eloafran@bitwise.io>