Load scar files with transact, not scabbard client
Update the scabbard CLI to use transact's `contract-archive` feature for
loading .scar files and remove the .scar loading capability from the
scabbard client (this is no longer necessary with transact's .scar
loading).

This changes the scabbard CLI's upload command to take the
`name:version` of the contract to load and optional `path`s.

Signed-off-by: Logan Seeley <seeley@bitwise.io>