Add package-lock.json to gitignore
These lock files shouldn't be committed with our current paradigm for
dependencies.

This is required due to a bug with yarn and git dependencies which cause
lifecycle scripts not to be run on yarn install. This means we have to
use npm install for splinter-canopyjs until it is published to the npm
registry.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>