Return a Result in PartialConfigBuilder.build()
Previously, all error handling was done in the constructors, requiring
that the build actually be fully done in the constructor. This isn't how
builders normally work, since one of the benefits of the builder pattern
is the ability to return errors that require full context of the items
set within the builder.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>