Insert into scabbard BatchHistory on update
Updates the scabbard `BatchHistory` struct, which is used to track batch
statuses, to insert batch updates if the batches are not already
present. Previously, if a batch wasn't pending, any updates would be
ignored; this meant that nodes that didn't receive the batch locally
(and instead received it over the network) would not have the batch's
status in their history.

Signed-off-by: Logan Seeley <seeley@bitwise.io>