Add build.rs file for generating man pages
The build.rs file will take the markdown files in
cli/man and convert them to man pages stored in
packaging/man. Note that if pandoc is not installed
the build.rs file will fail. To turn off manpage
generation set the environment variable
SPLINTER_NO_MAN_PAGES=true. If this environment
variable does not exist or is set to false, the man
pages will be generated.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>