Implement retry reconnection for disconnected connections
When a connection is discovered to be disconnected, the connection
manager will try to reconnect to the connection. If the
reconnection is not successful, the connection manager will wait
(with an exponential back off) until reconnection is successful.
The max retry frequency is configurable and defaults to 300 seconds.

This commit also simplifies ConnectionManagerNotification to
only include Connected and Disconnected. These messages will
be sent to subscribers when connection and reconnection is
successful. Disconnected is sent if a heartbeat fails and
reconnection is started.

This commit also makes the heartbeat interval configurable.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>