Relax pandoc file checks in cli build
While searching the PATH for pandoc, relax the error checking when
encountering an error listing a directory or reading an entry in the
directory.  This prevents any issues that may occur if a directory
cannot be read (e.g. it is an old entry in the path that may no longer
exist).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>