Add ResponseKey object
Adds a ResponseKey object to be used by key management routes that
are returning the internal Key structure, to allow for separate
control of what is returned in the REST endpoint responses.

Also implements From to translate the internal representation of a
Key into the ResponseKey REST object and converts Key objectst that
are currently being returned by REST endpoints into the ResponseKey
object.

Also removes the Serialize derive from the internal Key object, as
this is no longer necessary since it is not being used in the REST
API.

Signed-off-by: Shannyn Telander <telander@bitwise.io>