Add response body to key management post route
Adds a json body to the key management post route to add a key to
the KeyStore. This response body returns the resource that has been
created.

Also adds the Clone derive trait to the internal Key object to allow
for references to the values of the Key to be returned in the body.

Also fixes the openapi spec to accurately reflect the response body.

Signed-off-by: Shannyn Telander <telander@bitwise.io>