Add ability to set active signing keys
Adds the ability to set which pair of signing keys are currently active
for the user in this Canopy context. The user may be asked to provide
the password used to encrypt the private key if the session has timed
out. This also adds the ability to update the display name of keys.

Signed-off-by: Davey Newhall <newhall@bitwise.io>