Implement versioned frames for socket connections
A versioned frame is used for reading and writing data over socket
connections (or any binary Read/Write stream, for that matter).

The version for a connection is established via small negotiation
exchange, where the (min, max) pair of version are sent by the
outbound (local) stream, and the compatible version is sent by the
inbound (remote) stream.

Frame versions have an associated FrameHeader.  This header specifies
things about its version, the length of the frame as well as other
details about the frame.  The V1 header specifies length and a checksum
of the header, to ensure that the data received is the correct length,
and the length value is not a random value from anywhere along the
stream.

Frames themselves are containers for the data.  These are either owned
(for received data), or borrowed (for sent data). Only the sent data
needs to know the outbound version.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>