Add BindError for failed RestApi binding
Add the `RestApiServerError::BindError` enum variant and update the
`RestApi::run` method to report this error when it fails to bind to its
given URL. Previously, the calling code had no way of determining if an
error from the `run` method was due to an issue with the chosen bind URL
or something else. This allows the calling code to pick a different bind
URL (likely a different port) if binding fails.

Signed-off-by: Logan Seeley <seeley@bitwise.io>