Remove CmResponse from Connector public api
CmMessage is an internal message that is used to send messages from a
client (via the Connector) to the background thread.  With this change,
the connector's methods return idiomatic Rust responses, such as results
and options. This simplifies the use of the connectors function at the
call site.

To facilitate this change, each request contains a sender that is
specific to the response returned. As a result, CmRequest and CmPayload
have been merged into a single CmRequest enum.

The CmMessage and its associated structs have been moved into the mod
and have been made private.  These messages are internal only.

The Pacemaker has been modified to be message agnostic, and does not
need to know about the internal messages being sent.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>