Update admin service endpoint documentation
Add endpoint documentation to the `RestResourceProvider` implementations
of `AdminService` and `CircuitResourceProvider`. This documentation
mentions that one of the REST API backend features must be enabled for
these endpoints to be available, even though `rest-api-actix` is the
only backend currently.

Even though some of the endpoints are guarded by the `circuit-read` and
`proposal-read` features, these are not called out in the documentation
because these features will likely be removed when they are stabilized.

Signed-off-by: Logan Seeley <seeley@bitwise.io>