Move template service ID generation to base62 mod
The functions responsible for incrementing the service IDs in the
circuit templates feature are moved to the base62 module so they can be
used by other code that may need to manipulate base62 strings (other
code may want to use the same approach for generating predictable IDs).

In addition to moving the code, the methods are renamed to be more
generic, and the implementations are simplified.

Signed-off-by: Logan Seeley <seeley@bitwise.io>