Verify service ID in SplinterServiceBuilder
The `SplinterServiceBuilder::build` method is updated to verify that the
specified service ID is of the correct format. The correct format for a
service ID is a 4 character base62 (ASCII alphanumeric) string.

Previously, service IDs could be any string. This commit is the first
time the new service ID format is enforced in splinter.

Signed-off-by: Logan Seeley <seeley@bitwise.io>