Verify circuit ID in CreateCircuitBuilder
The `CreateCircuitBuilder::build` method is updated to verify that the
specified circuit ID is of the correct format. The correct format for a
circuit ID is an 11 character string composed of two, 5 character base62
(ASCII alphanumeric) strings joined with a dash (`-`). Example:
abcDE-F0123

Previously, circuit IDs could be any string. This commit is the first
time the new circuit ID format is enforced in splinter.

Signed-off-by: Logan Seeley <seeley@bitwise.io>