Add a reference map to network
This is a reusable map that can be used for reference
counting. It will keep track of how many times an id
has been added to the map, and only remove an entry
if the reference count makes it to zero.

It also includes a redirect map, which allows a
references id to be updated, while still allowing
the old references to be included.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>