Fix bug in connection manager reconnect
This bug would cause the heartbeats to try to be sent
twice before reconnection. It would also try to send
a heartbeat to connection that have already been flagged
as reconnecting, instead of just trying to reconnect.

This commit moves the heartbeat so it only get sent if
the connection is not trying to reconnect.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>