Add Initial Implementation of the PeerManager
The PeerManager is the component of the system that will know
about peers. It has a connector to the connection manager and
will request the creation and removal of connection that
are associated with a peer. The PeerManager is also in charge
of keeping track of reference counts of the peers, as well as
the metadata about a peer, such as its id, a list of endpoints
and which endpoint is currently active.

The PeerManager can be communicated with by using the
PeerManagerConnnector. The connector allows the following
request:

- add_peer_ref - adds a new peer or adds a reference count for
	an existing one. Returns a PeerRef object that when
	dropped will automatically send a removal request to
	the peer_manger.
- update_peer_ref - Update the peer id of an already created
	peer to a new id.
- list_peers - get a list of all currently connected peers
- subscribe - Subscribe to notifications about peers. Returns
	a PeerNotificationIter that will send connected and
	disconnected messages about peers.

The PeerManager also monitors notifications from the
ConnectionManager. If a connection disconnects, the PeerManager
will wait for specified number of reconnection attempts and if
the connection does not reconnect the peers other endpoints will
be tried. If successful, the peers active endpoint is replaced.
There is only ever one connection per peer active at a time.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>