Store gameroom service ID in DB
Updates the way gameroom determines the service ID of the scabbard
instance to send batches to. Previously, gameroom assumed that the
service ID was 'gameroom_<NODE_ID>`, but that is no longer a valid
service ID and will be changed in another commit.

This update has gameroom store the service ID for a newly created
gameroom in the gameroom database, and fetch the service ID when a
request to the gameroom daemon's `POST /gamerooms/{circuit_id}/batches`
endpoint is made.

Additionally, this commit fixes some of the OpenaAPI documentation for
the `POST /gamerooms/{circuit_id}/batches` endpoint.

Signed-off-by: Logan Seeley <seeley@bitwise.io>