Add/update structs for proposal-read responses
Add new structs that define the format of the responses for the
`proposal-read` endpoints. These structs are defined in two separate
files, one for each endpoint (just like the `circuit-read` endpoints).

This change means the `proposal-read` endpoints do not return any
internally-defined splinter structures; this allows the internal
structures to change as needed without impacting the data that's exposed
via the REST API (and vice versa).

Signed-off-by: Logan Seeley <seeley@bitwise.io>