Update OpenAPI docs for circuit- and proposal-read
Updates the OpenAPI documentation for the responses of the
`circuit-read` and `proposal-read` endpoints, which have been updated.

For documentation purposes, both `circuit-read` endpoints share the same
definition of a `Circuit` response in the OpenAPI file, even though each
endpoint has their own definition in the code. The code definitions are
identical for now, so the single OpenAPI definition is used for
simplicity. If the responses of these endpoints differs in the future,
the documentation will need to be updated to have separate definitions.
The same goes for the `proposal-read` endpoints.

Signed-off-by: Logan Seeley <seeley@bitwise.io>