Save disconnected connection and return on removal in Mesh
Before if a connection disconnected, it was removed. Then in
the future if the connection was removed, an error was returned
because the connection did not exist.

Now when a connection is removed, it will be stored in a
HashMap for disconnected connection. If the connection is
removed at a later time, the connection will be removed
from this map and returned, as if the connection had
never disconnected.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>