Remove derive Clone from internal messages
Remove the derive of Clone for the internal message sent back and forth
between connectors and the background thread.  This allows for future
messages to contain data that is not cloneable.

This requirement was only necessary to support the Pacemaker's use of a
a message.  As a result of removing clone, the start method has been
modified to take a function that will create a message to be sent.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>