Add a BiHashMap to Mesh for a unique id to mesh id
This change enables components using mesh to not have
to know about the internal mesh id used in polling the
connections. This also allows reconnection with out having
to replace the id returned from mesh.

This commit updates mesh and all places mesh is used. It
also removes the use of rwlock unwrap macros and instead
returns a PoisonedLock error instead.

This will simplify the data future inteconnects will need
to keep track of.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>