Allow look-ups with Borrow implementations
Like HashMap, allow the BiHashMap to support get_by_* via elements that
K or V implement the Borrow trait on.  This removes the need, for
example, to clone strings in order to look up a string key.

For example:

    bi_map.get_by_key(&"some-key".to_string())

can now be simplified to

    bi_map.get_by_key("some-key")

Additionally, this changes contains_* and remove_by.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>