Update splinter readme to pull gridd build image
Updates the grid-on-splinter README doc to properly pull in the builder
image (hyperledger/grid-dev:v1) for gridd when building. Previously, the
instructions had the user pull in the `hyperledger/grid-dev:latest`
image, so the builder image that's used wasn't getting updated. Now the
user is instructed to build images using `docker-compose build --pull`,
which will pull in the base builder images on build.

Signed-off-by: Logan Seeley <seeley@bitwise.io>