Update splinter dependency to git master branch
Updates the `splinter` dependency for gridd to pull from the master
branch in the git repository rather than from crates.io. This change is
required to fix the grid-on-splinter example, since it currently uses
splinterd images built from the splinter master branch; this resulted in
a mismatch of splinter versions between gridd and splinterd.

When all splinter features required for the grid-on-splinter example
have been stabilized and released, this should be reverted back to
pulling from crates.io and the docker images should be updated to tagged
releases rather than `experimental`.

Signed-off-by: Logan Seeley <seeley@bitwise.io>