Update product insert db operations
Updates the product insert/update database operations to account for
service ID when adding or modifying a product. This prevents products
from being 'overwritten' when adding a product to multiple services.
Before the behavior would incorrectly handle incrementing the
`end_commit_num` for some product records. This also adds the
'Identifiable' trait to the Product and ProductPropertValue structs.
This was done to prevent Diesel from throwing an error when inserting
records with a non-standard name for the primary key ('id' is the
default). Along with this, the table name is now specified. Again, this
is to prevent Diesel from throwing errors when inserting into tables
with non-standard names ('product' instead of 'products').

Signed-off-by: Davey Newhall <newhall@bitwise.io>