Update redirects in login sapling
Update how the login sapling redirects a user to the page they were
attempting to navigate to. This is to make it compatible with updates
being made to Canopyjs which will now redirect users to the login page
if they're not logged in. Canopyjs redirects will populate a field in
`$CANOPY` called `redirected_from`, so login simply needs to return that
user to that page after they have been logged in.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>