Fix sapling tab rendering on refresh
Fixes an issue in Canopy UIs that display saplings in nav tabs would
disappear and then quickly reappear on refresh. This was because the
sapling state was being initialized as an empty array.

This commit fixes the issue by caching the sapling state in session
storage, so the state is initialized as the previously fetched value.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>