Store filter/sort functions in state
Instead of storing the filtered values for each type of filter or the
sort functions, store the functions that will be applied.  This
simplifies the application of the functions on the circuit lists.

Additionally, this changes allows for the circuit list to be refreshed
without removing the filter or sort order.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>