Replace react scripts with webpack
Replace the use of react-scripts with the use of a simple webpack
configuration.  React scripts are very opinionated and have no real way
to configure them without using the rewire library.  Likewise, it adds a
lot of complexity to the build, which makes for a slow development
cycle.

This change configures webpack for production mode for build and deploy.
It switches to development mode for bundle and deploy-local.  This
enables debugging the saplings, even when loaded by canopy.

The start script is removed, as saplings (currently) do not have a way
to run in isolation from canopy.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>