Resolve ./src as App
Resolve sources at the root source level, as "App". This allows for
components to be referred by an absolute path, which makes refactoring
easier.

Instead of

    const { Circuit } = require('../../data/circuit');

where the paths are relative to the nesting of the current file, the
requirement can be specified as

    const { Circuit } = require('App/data/circuit');

This path will work regardless of where the current file is located.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>