Replace ServiceCard with ServiceTable/-Form
This change replaces the ServiceCard with a more readable combination of
a ServiceTable and a ServiceForm.

The ServiceTable can be used in both edit and read-only modes.  The
former adds a set of controls that will emit onEdit and onDelete events.

The ServiceForm is standard form for adding services entries to the
table.  It is configured to use the Overlay panel in propose circuit
form during the service step.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>