Add a complete Dockerfile for admin UI
Add a single Dockerfile that will serve both the canopy framework, and
the three saplings available for this repository.  This Dockerfile
should result in a image that is publishable to Docker Hub.

The Dockerfile exposes one arg: REPO_VERSION and expects one environment
variable: SPLINTER_URL.

Note, that this commit enables CORS headers for the saplings, as it
doubles as a sapling repository.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>