Refactor DB and commit functionality
This refactors the Grid DB to follow the pattern established in the
Cargill/splinter project. This refactor consists of moving the DB
functionality primarily to the Grid SDK and moving the operation
handlers for the commits DB module. The database handler remains in the
Grid daemon because that is where events are actually received. This
first PR only moves the operations and database functionality for the
commits module but other modules can be progressively added/refactored
following the pattern of the commits module. Additionally, database
functionality is behind the "database" feature flag in the SDK.

Signed-off-by: Davey Newhall <newhall@bitwise.io>