Check if product gs1 schema is present
When creating or updating a product, if the namespace is set to GS1, the
product smart contract will now check to make sure the product meets the
requirements of the schema. This includes checking that all required
fields are set and that there are no fields included in the product that
are not defined by the schema.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>