Reorganize Grid UI directory and update Dockefiles
This commit reorganizes the Grid UI directory and Dockerfiles to match
the format in the splinter-ui repo. There is now a top level /ui dir
that contains all of the Grid UI related files. Under that, the /grid-ui
directory has the Canopy UI for grid, /saplings has all of the Grid
sapling source, and /sapling-dev-server has the sapling config for the
Grid UI. The sapling-dev-server was moved from its own Docker container
to the grid-ui container.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>