Update product commands
This updates the product create, update, and delete commands. The create
and update commands can now add a single product using command line args
to pass required information. Additionally, the `--path` arg is switched
to `--file` and can now take either a file path or a URL. The accepted
format of the YAML file is also slightly different with properties now
taking the form of `<property_name>: <property_value>` rather than
having to specify the name, type, and value separately for each
property.

Signed-off-by: Davey Newhall <newhall@bitwise.io>