Add protos to UI directory and add to build
The protos have to be added to the ui directory since it is used as the
context directory for Grid UI builds. This also fixes the protogen
script to use these files.

This commit also fixes the gitignore since it was ignoring the artifact
of the generate proto files script twice.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>