Use webpack for product sapling build
Switches the react-scripts CRA pipeline to a vanilla webpack one. This
allows us to build the sapling in dev mode and see any errors in the
console. We were not using the features of the CRA pipeline for saplings
so it does not make sense to use it over the more flexible webpack
pipeline.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>