Have CLI correctly communicate connection errors
Adds a new CLI error type called DaemonError which is meant to
communicate any issues talking to the grid daemon. Furthermore, the
parts of the CLI that sent requests to the grid daemon will now interpret
non-2xx status codes as fatal errors, and neatly print that output to standard
out.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>