Remove the phrase "Reqwest Error" from ReqwestError display
Reqwest Error is meant to refer to errors emmitted by the http library,
Reqwest, but since most of these errors are printed to standard out, it
makes it difficult to parse those errors for the user, since the user
may not know the name of the http library used by the CLI, nor should
they. As such, we are removing that information from the display.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>