Overhaul Grid REST API spec
This commit is a heavy edit of the Grid REST API specification. The
specification was outdated and had many structural and consistency
issues. Due to the scope of changes, the spec was rewritten from scratch
but pulls in some models and structure from the previous version as
appropriate.

In summary, the following changes were made:
 - Updated the info object
 - Added documentation and links to transaction routes
 - Standardized summary and description format
 - Removed inaccurate error types
 - Organized schemas and paths by feature

Signed-off-by: Darian Plumb <dplumb@bitwise.io>